Split kernels.py into _kernels.py and the imports.
