--- statsmodels/kernel_methods/_kernelsnd.py
@@ -1,7 +1,0 @@
-"""
-Module containing kernels in nD, except for Gaussian, which is already in
-the main kernels module.
-
-Note: There are no such kernels for now.
-"""
-#from .kernels import KernelnD
